+++ lib/service_actor/success.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module ServiceActor
-  # Raised when using `succeed!` to halt the progression of an organizer.
-  # DEPRECATED in favor of adding conditions to your play.
-  class Success < Error; end
-end