+++ spec/examples/disallow_nil_on_input_with_deprecated_required.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class DisallowNilOnInputWithDeprecatedRequired < Actor
-  input :name, type: String, required: true
-
-  def call; end
-end